feat(): restructure, add PDUType back to every PDU implementation, fix chat history not clearing after leaving a lobby and available lobbies not showing after connecting to the server
Took 3 hours 37 minutes
