+++ src/main/java/server/game/docker/net/modules/messages/decoders/PDUChatMessageDecoder.java
@@ -1,0 +1,25 @@
+package server.game.docker.net.modules.messages.decoders;
+
+import io.netty.buffer.ByteBuf;
+import io.netty.channel.Channel;
+import server.game.docker.net.modules.messages.pdus.PDUChatMessage;
+import server.game.docker.net.parents.decoders.PDUHandlerDecoder;
+import server.game.docker.net.parents.handlers.PDUInboundHandler;
+
+import java.nio.charset.Charset;
+
+public class PDUChatMessageDecoder implements PDUHandlerDecoder {
+    private static final int AUTHOR_NAME_LENGTH = 8;
+    private static final int MAX_MESSAGE_LENGTH = 64;
+
+    @Override
+    public void decode(ByteBuf in, Channel channel, PDUInboundHandler out) {
+        PDUChatMessage chatMessage = new PDUChatMessage();
+
+        chatMessage.setAuthorID(in.readLong());
+        chatMessage.setAuthorName(in.toString(in.readerIndex(), AUTHOR_NAME_LENGTH, Charset.defaultCharset()).trim());
+        chatMessage.setMessage(in.toString(in.readerIndex() + AUTHOR_NAME_LENGTH, MAX_MESSAGE_LENGTH, Charset.defaultCharset()).trim());
+
+        out.handle(chatMessage, channel);
+    }
+}
