+++ src/main/java/server/game/docker/net/parents/handlers/PDUHandler.java
@@ -3,5 +3,6 @@ package server.game.docker.net.parents.handlers;
 /**
  * Tagging interface for classes representing a PDU handler.
  */
+@Deprecated
 public interface PDUHandler {
 }
