+++ src/main/java/server/game/docker/net/parents/routers/PDURouter.java
@@ -1,0 +1,4 @@
+package server.game.docker.net.parents.routers;
+
+public interface PDURouter {
+}
