--- src/main/java/server/game/docker/server/persistence/modules/MatchPlayer.java
@@ -1,6 +1,0 @@
-package server.game.docker.server.persistence.modules;
-
-public class MatchPlayer {
-    private Long iD;
-    private Long clientID;
-}
