+++ src/main/java/server/game/docker/server/persistence/modules/matches/MatchPlayer.java
@@ -1,0 +1,6 @@
+package server.game.docker.server.persistence.modules.matches;
+
+public class MatchPlayer {
+    private Long iD;
+    private Long clientID;
+}
